+++ src/main/java/com/dailyon/productservice/common/feign/request/OrderProductDto.java
@@ -1,0 +1,15 @@
+package com.dailyon.productservice.common.feign.request;
+
+import lombok.Builder;
+import lombok.Getter;
+
+import javax.validation.constraints.NotNull;
+
+@Getter
+@Builder
+public class OrderProductDto {
+    @NotNull
+    private Long productId;
+    @NotNull
+    private Long sizeId;
+}
