+++ src/main/java/com/dailyon/productservice/productstock/repository/ProductStockCustomRepository.java
@@ -1,6 +1,6 @@
 package com.dailyon.productservice.productstock.repository;
 
-import com.dailyon.productservice.product.dto.request.ReadOrderProductRequest;
+import com.dailyon.productservice.common.feign.request.OrderProductDto;
 import com.dailyon.productservice.productstock.entity.ProductStock;
 
 import java.util.List;
@@ -8,5 +8,5 @@ import java.util.List;
 
 public interface ProductStockCustomRepository {
 
-    List<ProductStock> findOrderProductsBy(List<ReadOrderProductRequest.ProductDto> productDtos);
+    List<ProductStock> findOrderProductsBy(List<OrderProductDto> productDtos);
 }
