+++ src/main/java/com/dailyon/productservice/productstock/repository/ProductStockCustomRepositoryImpl.java
@@ -1,6 +1,6 @@
 package com.dailyon.productservice.productstock.repository;
 
-import com.dailyon.productservice.product.dto.request.ReadOrderProductRequest;
+import com.dailyon.productservice.common.feign.request.OrderProductDto;
 import com.dailyon.productservice.productstock.entity.ProductStock;
 import com.querydsl.core.BooleanBuilder;
 import com.querydsl.jpa.impl.JPAQueryFactory;
@@ -12,7 +12,6 @@ import java.util.List;
 import static com.dailyon.productservice.product.entity.QProduct.product;
 import static com.dailyon.productservice.productsize.entity.QProductSize.productSize;
 import static com.dailyon.productservice.productstock.entity.QProductStock.productStock;
-import static com.dailyon.productservice.category.entity.QCategory.category;
 
 
 @Repository
@@ -21,9 +20,9 @@ public class ProductStockCustomRepositoryImpl implements ProductStockCustomRepos
     private final JPAQueryFactory jpaQueryFactory;
 
     @Override
-    public List<ProductStock> findOrderProductsBy(List<ReadOrderProductRequest.ProductDto> productDtos) {
+    public List<ProductStock> findOrderProductsBy(List<OrderProductDto> productDtos) {
         BooleanBuilder builder = new BooleanBuilder();
-        for(ReadOrderProductRequest.ProductDto dto: productDtos) {
+        for(OrderProductDto dto: productDtos) {
             builder.or(
                     (productStock.product.id.eq(dto.getProductId())
                         .and(productStock.productSize.id.eq(dto.getSizeId()))
