+++ src/main/java/com/dailyon/productservice/productstock/service/ProductStockService.java
@@ -1,7 +1,7 @@
 package com.dailyon.productservice.productstock.service;
 
-import com.dailyon.productservice.product.dto.request.ReadOrderProductRequest;
-import com.dailyon.productservice.product.dto.response.ReadOrderProductListResponse;
+import com.dailyon.productservice.common.feign.request.OrderProductDto;
+import com.dailyon.productservice.common.feign.response.ReadOrderProductListResponse;
 import com.dailyon.productservice.productstock.repository.ProductStockRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -15,8 +15,7 @@ import java.util.List;
 public class ProductStockService {
     private final ProductStockRepository productStockRepository;
 
-    public ReadOrderProductListResponse readOrderProducts(ReadOrderProductRequest request) {
-        List<ReadOrderProductRequest.ProductDto> dtoList = request.getProductRequest();
-        return ReadOrderProductListResponse.fromEntity(productStockRepository.findOrderProductsBy(dtoList));
+    public ReadOrderProductListResponse readOrderProducts(List<OrderProductDto> request) {
+        return ReadOrderProductListResponse.fromEntity(productStockRepository.findOrderProductsBy(request));
     }
 }
